+++ ganttzk/src/main/resources/web/js/ganttz/TaskRow.js
@@ -1,21 +1,27 @@
 zk.$package("ganttz");
 
 ganttz.TaskRow = zk.$extends(zk.Widget, {
+
     _labelsHidden : true,
-    showLabels : function(){
+
+    showLabels : function() {
         this._labelsHidden = false;
         this.firstChild.showLabels();
     },
-    hideLabels : function(){
+
+    hideLabels : function() {
         this._labelsHidden = true;
         this.firstChild.hideLabels();
     },
+
     _resourcesHidden : true,
-    hideResourceTooltip : function(){
+
+    hideResourceTooltip : function() {
         this._resourcesHidden = true;
         this.firstChild.hideResourceTooltip();
     },
-    showResourceTooltip : function(){
+
+    showResourceTooltip : function() {
         this._resourcesHidden = false;
         this.firstChild.showResourceTooltip();
     }